+++ projects/TreeSketch/src/TreeSketchTreeObj.cpp
@@ -6,13 +6,13 @@
 namespace zeno
 {
     TreeObj::BranchObj::BranchObj(
-        const zeno::vec4d &start, const zeno::vec4d &direction, const double length, const double radius,
+        const zeno::vec4d &start, const zeno::vec4d &direction, const float length, const float radius,
         const std::vector<zeno::vec4d> &turn_points)
         : _start{start}, _direction{direction}, _length{length}, _radius{radius},
           _turn_points{turn_points}, _hasLeaf{false}, _children{} {}
 
     zeno::vec4d TreeObj::BranchObj::calculate_child_start(
-        const double offset_start_min, const double offset_start_max)
+        const float offset_start_min, const float offset_start_max)
     {
         auto offset_start{random(offset_start_min, offset_start_max)};
 
@@ -40,18 +40,18 @@ namespace zeno
             high_point = _turn_points[high];
         }
 
-        auto t{mid - static_cast<double>(low)};
+        auto t{mid - static_cast<float>(low)};
 
         return lerp(low_point, high_point, t);
     }
 
     TreeObj::TreeObj(
         const zeno::vec4d &start,
-        const double offset_radian_min, const double offset_radian_max,
-        const double length_min, const double length_max,
-        const double radius_min, const double radius_max,
+        const float offset_radian_min, const float offset_radian_max,
+        const float length_min, const float length_max,
+        const float radius_min, const float radius_max,
         const int turn_points_num_min, const int turn_points_num_max,
-        const double turn_points_offset_min, const double turn_points_offset_max)
+        const float turn_points_offset_min, const float turn_points_offset_max)
         : _tree_level{0}
     {
         std::srand(std::time(nullptr));
@@ -74,12 +74,12 @@ namespace zeno
 
     void TreeObj::create_branchs(
         const int num_min, const int num_max,
-        const double offset_start_min, const double offset_start_max,
-        const double offset_radian_min, const double offset_radian_max,
-        const double length_min, const double length_max,
-        const double radius_min, const double radius_max,
+        const float offset_start_min, const float offset_start_max,
+        const float offset_radian_min, const float offset_radian_max,
+        const float length_min, const float length_max,
+        const float radius_min, const float radius_max,
         const int turn_points_num_min, const int turn_points_num_max,
-        const double turn_points_offset_min, const double turn_points_offset_max)
+        const float turn_points_offset_min, const float turn_points_offset_max)
     {
         std::queue<BranchObj *> branch_queue;
         std::queue<int> level_queue;
